Moved profiler callback to its own file and added some more settings.
This should make it easier to control the program for development work and to enable/disable the profiler logging.
