+++ src/settings.ts.template
@@ -1,1 +1,3 @@
 export const coniqlSocket = "localhost:1000";
+export const simulationTime = 100;
+export const profilerEnabled = true;
